+++ controladores/controlador.head.php
@@ -3,18 +3,11 @@
 
 class ControladorHead{
     
-    // IMPORTA EL HEAD A LAS PAGE
-    public function head(){
-
-        include "vistas/modulos/head.modelos.php";
-
-    }
-
 
     // IMPORTA EL HEAD PARA LOGIN
-    public function headLogin(){
+    public function HeadModelos(){
 
-        include "vistas/modulos/head.login.php";
+        include "vistas/modulos/head.modelos.php";
 
     }
