+++ vistas/js/principal.js
@@ -9,8 +9,8 @@ $(document).ready(function(){
         // Store hash
         var hash = this.hash;
   
-        // Using jQuery's animate() method to add smooth page scroll
-        // The optional number (900) specifies the number of milliseconds it takes to scroll to the specified area
+        // ?Using jQuery's animate() method to add smooth page scroll
+        // !The optional number (900) specifies the number of milliseconds it takes to scroll to the specified area
         $('html, body').animate({
           scrollTop: $(hash).offset().top
         }, 900, function(){
